worker/requirements.txt
python-dotenv
motor
ccxt
sendgrid
twilio